Import new discovery, meta info in diffHistory.md
