+++ packages/backend/package.json
@@ -35,7 +35,7 @@
     "@koa/router": "^12.0.0",
     "@l2beat/backend-tools": "^0.5.0",
     "@l2beat/config": "*",
-    "@l2beat/discovery": "0.43.7",
+    "@l2beat/discovery": "0.44.0",
     "@l2beat/discovery-types": "0.8.0",
     "@l2beat/shared": "*",
     "@l2beat/shared-pure": "*",
